# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Skoll API
  version: 1.0.0
  description: OpenAPI document for Skoll core HTTP endpoints.
servers:
  - url: __API_BASE_PREFIX__
paths:
...
  /v1/plugins/pharma_oa/api/customers:
    get:
      summary: List Pharma OA customers
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, disabled]
        - name: region
          in: query
          schema:
            type: string
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, default: 0}
        - name: cursor
          in: query
          description: Opaque cursor returned by the previous page. When present, it takes precedence over offset.
          schema: {type: string}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PharmaMasterListAPIResponse'}
        '403':
          description: Forbidden
    post:
      summary: Create Pharma OA customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name, region, organizationId, ownerId]
              properties:
                code:
                  type: string
                name:
                  type: string
                region:
                  type: string
                organizationId:
                  type: string
                ownerId:
                  type: string
                rating:
                  type: integer
                  minimum: 0
                  maximum: 5
                contacts:
                  type: array
                  items:
                    type: object
                qualifications:
                  type: array
                  items:
                    type: object
      responses:
        '201':
          description: Created
        '403':
          description: Forbidden
  /v1/plugins/pharma_oa/api/customers/{id}:
    put:
      summary: Update Pharma OA customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
  /v1/plugins/pharma_oa/api/customers/{id}/disable:
    post:
      summary: Disable Pharma OA customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
  /v1/plugins/pharma_oa/api/customers/{id}/sales-eligibility:
    get:
      summary: Validate Pharma OA customer sales eligibility
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
  /v1/plugins/pharma_oa/api/customers/qualification-reminders:
    get:
      summary: List Pharma OA customer qualification reminders
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
...
  /v1/plugins/pharma_oa/api/demo-seed/status:
    get:
      summary: Read Pharma OA demo seed status
      responses:
        '200':
          description: Current demo seed state and entity references
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DemoSeedAPIResponse'}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '403':
          description: Missing demo seed read permission
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/plugins/pharma_oa/api/demo-seed/apply:
    post:
      summary: Initialize the complete Pharma OA demo scenario
      responses:
        '200':
          description: Demo scenario initialized or existing result reused
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DemoSeedAPIResponse'}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '403':
          description: Missing demo seed apply permission
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '500':
          description: Demo seed failed and exposes its terminal stage through the status operation
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Skoll JWT access token. Aggregated plugin API operations require this scheme and declare x-skoll-permission.
  schemas:
...
    PharmaMasterListData:
      type: object
      required: [items, offset, limit, total, hasMore, nextCursor, sort]
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        offset: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1, maximum: 200}
        total: {type: integer, format: int64, minimum: 0}
        hasMore: {type: boolean}
        nextCursor:
          type: string
          description: Empty on the final page; otherwise pass this value as the next request cursor.
        sort:
          type: string
          enum: ['code:asc']
    PharmaMasterListAPIResponse:
      allOf:
        - $ref: '#/components/schemas/MessageResponse'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/PharmaMasterListData'
...
    DemoSeedEntities:
      type: object
      required: [employeeId, productId, supplierId, customerId, warehouseId, purchaseRequestId, workflowInstanceId, purchaseOrderId, purchaseInboundId, batchId, salesOrderId, salesOutboundId, customerFollowUpId]
      properties:
        employeeId: {type: string}
        productId: {type: string}
        supplierId: {type: string}
        customerId: {type: string}
        warehouseId: {type: string}
        purchaseRequestId: {type: string}
        workflowInstanceId: {type: string}
        purchaseOrderId: {type: string}
        purchaseInboundId: {type: string}
        batchId: {type: string}
        salesOrderId: {type: string}
        salesOutboundId: {type: string}
        customerFollowUpId: {type: string}
    DemoSeedSnapshot:
      type: object
      required: [state, stage, applied, reused, entities, counts]
      properties:
        state: {type: string, enum: [ready, applying, applied, failed]}
        stage: {type: string}
        applied: {type: boolean}
        reused: {type: boolean}
        actorId: {type: string}
        appliedAt: {type: string, format: date-time}
        error: {type: string}
        entities: {$ref: '#/components/schemas/DemoSeedEntities'}
        counts:
          type: object
          additionalProperties: {type: integer, minimum: 0}
    DemoSeedData:
      type: object
      required: [item]
      properties:
        item: {$ref: '#/components/schemas/DemoSeedSnapshot'}
    DemoSeedAPIResponse:
      allOf:
        - {$ref: '#/components/schemas/MessageResponse'}
        - type: object
          properties:
            data: {$ref: '#/components/schemas/DemoSeedData'}
...
    MessageResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: error
        message:
          type: string
...